LPS-140383 Rename Activation Keys Table Component to reuse in Portal Page
