+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/TableKeyDetails/index.js
@@ -19,7 +19,7 @@ import {
 	getProductName,
 	getStatusActivationTag,
 	hasVirtualCluster,
-} from '../../containers/DXPActivationKeysTable/utils';
+} from '../../containers/ActivationKeysTable/utils';
 
 const HOST_NAME = 'Host Name';
 const IP_ADDRESSES = 'IP Addresses';
