+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/pages/Project/DXP/index.js
@@ -8,12 +8,12 @@
  * permissions and limitations under the License, including but not limited to
  * distribution rights of the Software.
  */
-import DXPActivationKeysTable from '../../../containers/DXPActivationKeysTable';
+import ActivationKeysTable from '../../../containers/ActivationKeysTable';
 import DeveloperKeysLayouts from '../../../layouts/DeveloperKeysLayout';
 const DXP = ({project, sessionId}) => {
 	return (
 		<div className="mr-4">
-			<DXPActivationKeysTable project={project} sessionId={sessionId} />
+			<ActivationKeysTable project={project} sessionId={sessionId} />
 
 			<DeveloperKeysLayouts>
 				<DeveloperKeysLayouts.Inputs
