+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/styles/_containers.scss
@@ -1,3 +1,3 @@
-@import 'containers/DXPActivationKeysTable/dxp-activation-keys-table';
+@import 'containers/ActivationKeysTable/activation-keys-table';
 @import 'containers/SideMenu/side-menu';
 @import 'containers/TeamMembersTable/team-members-table';
