+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/styles/containers/ActivationKeysTable/Header/_activation-keys-table-header.scss
