+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/styles/containers/ActivationKeysTable/_activation-keys-table.scss
@@ -1,4 +1,4 @@
-@import './Header/dxp-activation-keys-table-header';
+@import './Header/activation-keys-table-header';
 
 .cp-dxp-activation-key-container {
 	max-width: 1271px;
